% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_flatten_json.R
\name{bq_flatten_json}
\alias{bq_flatten_json}
\title{Takes a JSON column in a BigQuery table and produces another column of the same length.}
\usage{
bq_flatten_json(
  project_name,
  dataset_name,
  table_name,
  json_column_name,
  destination_table_name = NULL
)
}
\arguments{
\item{project_name}{A string of the BQ project name.}

\item{dataset_name}{A string of the name of the dataset where table lives.}

\item{table_name}{A string of the name of the data table.}

\item{json_column_name}{A string of the name of the table column that contains the JSON values.}

\item{destination_table}{A table in BQ to send the results to.}
}
\value{
A \code{\link[bigrquery:bq_table]{bigrquery::bq_table()}} that has a column of the flattened JSON strings.
}
\description{
Takes a JSON column in a BigQuery table and produces another column of the same length.
}
\examples{
bq_flatten_json("ultra-project","TEST_DATASET_3","test_table","string_field_0", "test_destination_table")
}
\seealso{
\code{\link[bigrquery:bq_dataset_query]{bigrquery::bq_dataset_query()}}, which this function wraps around.
}
