% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_get_fieldnames.R
\name{bq_get_json_fieldnames}
\alias{bq_get_json_fieldnames}
\title{Gets a list of first-level field names of all JSON values in a BQ column}
\usage{
bq_get_json_fieldnames(
  project_name,
  dataset_name,
  table_name,
  json_column_name
)
}
\arguments{
\item{project_name}{A string of the BQ project name}

\item{dataset_name}{A string of the name of the dataset where table lives}

\item{table_name}{A string of the name of the data table}

\item{json_column_name}{A string of the name of the table column that contains the JSON values}
}
\value{
A dataframe listing out the first-level field names of the JSONs in the specified column; this is the union of all keys used in any JSON there.
}
\description{
Gets a list of first-level field names of all JSON values in a BQ column
}
\examples{
bq_get_json_fieldnames("ultra-project","TEST_DATASET_3","test_demo_flattened","flattened_json")
}
\seealso{
\code{\link[bigrquery:bq_dataset_query]{bigrquery::bq_dataset_query()}}, which this function wraps around.
}
