% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_replace_string.R
\name{bq_replace_string}
\alias{bq_replace_string}
\title{Replaces a specified substring in the table, across specified columns}
\usage{
bq_replace_string(
  column_names,
  str_to_replace,
  new_string,
  project_name,
  dataset_name,
  table_name,
  destination_table_name = NULL
)
}
\arguments{
\item{column_names}{String that will be the replacement.}

\item{str_to_replace}{String that will be replaced.}

\item{project_name}{A string of the BQ project name.}

\item{dataset_name}{A string of the name of the dataset where table lives.}

\item{table_name}{A string of the name of the data table.}

\item{destination_table}{A table in BQ to send the results to.}
}
\value{
A \code{\link[bigrquery:bq_table]{bigrquery::bq_table()}} that has each field of the JSON values spread out as individual columns.
}
\description{
Replaces a specified substring in the table, across specified columns
}
\examples{
bq_get_fieldnames(json_fieldnames, "ultra-project","TEST_DATASET_3","test_table", "test_destination_table")
}
\seealso{
\code{\link[bigrquery:bq_dataset_query]{bigrquery::bq_dataset_query()}}, which this function wraps around.
}
